examples[model-loder]: fix CPU + shader byte size bug
Issues occur if the amount of bytes of data we use
on the CPU side is less than what the shader/GPU
requires. Change uint8_t to int so we the
CPU struct and the shader uniform struct variable
are the same exact size.

+++ examples/wayland/model-load.c
@@ -110,7 +110,7 @@ struct app_vertex_data {
 
 struct app_uniform_buffer_scene_model {
 	mat4 model;
-	uint8_t textureIndex;
+	int textureIndex;
 };
