+++ examples/xcb/model-load.c
@@ -105,7 +105,7 @@ struct app_vertex_data {
 
 struct app_uniform_buffer_scene_model {
 	mat4 model;
-	uint8_t textureIndex;
+	int textureIndex;
 };
